integrated summery view to home page

+++ lib/dto/mothly_transaction_summary_view.dart
@@ -2,8 +2,8 @@ import 'package:spend_wise/dto/transaction.dart';
 
 class MonthlyTransactionSummary {
   final List<TransactionDto> trasactions;
-  final int totalIncome;
-  final int totalExpense;
+  final double totalIncome;
+  final double totalExpense;
   final Map<String, int> expensesMap;
 
   MonthlyTransactionSummary(
